refactor(devtools): improve profiler visualizations
Improve the profiler's visualization modes visually.
