+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/recording-visualizer/bargraph-visualizer/bar-chart/bar-chart.component.scss
@@ -1,35 +1,38 @@
-.bar {
-  width: 0%;
-  margin-bottom: 7px;
-  align-items: center;
-  display: flex;
-  padding-top: 3px;
-  padding-bottom: 3px;
-  opacity: 0.8;
-  cursor: pointer;
-  transition:
-    opacity 0.3s ease-out,
-    width 0.3s ease;
-}
-
-.bar span {
-  margin-left: 10px;
-  max-width: calc(100% - 20px);
-  display: inline-block;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  opacity: 1;
-  color: var(--color-text);
-}
-
-.bar:hover {
-  opacity: 1;
-}
-
-:host,
-.wrapper {
-  width: calc(100% - 5px);
-  height: 100%;
+:host {
   display: block;
+
+  .wrapper {
+    padding-block: 0.75rem;
+    height: inherit;
+    box-sizing: border-box;
+
+    .bar {
+      width: 0;
+      margin-bottom: 0.25rem;
+      display: flex;
+      align-items: center;
+      padding-block: 0.2rem;
+      background-color: var(--dynamic-blue-02);
+      cursor: pointer;
+      transition:
+        background-color 0.3s ease-out,
+        width 0.3s ease;
+
+      &:hover {
+        background-color: color-mix(in srgb, var(--dynamic-blue-02) 70%, var(--full-contrast) 30%);
+        opacity: 1;
+      }
+
+      span {
+        padding-inline: 0.5rem;
+        max-width: 100%;
+        display: inline-block;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+        overflow: hidden;
+        opacity: 1;
+        color: var(--septenary-contrast);
+      }
+    }
+  }
 }
