+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/recording-visualizer/bargraph-visualizer/bargraph-visualizer.component.html
@@ -1,4 +1,1 @@
-<div class="level-profile-wrapper">
-  <ng-bar-chart (barClick)="selectNode($event)" [color]="barColor()" [data]="profileRecords()">
-  </ng-bar-chart>
-</div>
+<ng-bar-chart (barClick)="selectNode($event)" [data]="profileRecords()" />
