+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/recording-visualizer/bargraph-visualizer/bargraph-visualizer.component.scss
@@ -1,15 +1,6 @@
 :host {
-  width: 100%;
   height: 100%;
-  display: flex;
-  -webkit-user-select: text;
-  -ms-user-select: text;
   user-select: text;
   overflow-y: auto;
-}
-
-.level-profile-wrapper {
-  height: 100%;
-  width: 100%;
-  cursor: pointer;
+  display: block;
 }
