+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/recording-visualizer/execution-details/execution-details.component.scss
@@ -1,12 +1,21 @@
-table {
-  border-collapse: collapse;
-  width: 100%;
-  text-align: left;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
+:host {
+  display: block;
+  white-space: nowrap;
 
-th,
-td {
-  border-bottom: 1px solid var(--senary-contrast);
+  table {
+    border-collapse: collapse;
+    width: 100%;
+    text-align: left;
+
+    th,
+    td {
+      border-bottom: 1px solid var(--senary-contrast);
+      padding: 0.375rem 0.5rem;
+      overflow: hidden;
+      text-overflow: ellipsis;
+    }
+
+    .name {
+    }
+  }
 }
