+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/recording-visualizer/flamegraph-visualizer/flamegraph-visualizer.component.scss
@@ -1,32 +1,28 @@
 :host {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  -webkit-user-select: text;
-  -ms-user-select: text;
+  display: block;
   user-select: text;
+  overflow-y: auto;
+  overflow-x: hidden;
+  height: 100%;
 
-  .level-profile-wrapper {
-    height: 100%;
-    width: 100%;
-    cursor: pointer;
-    overflow-y: auto;
-  }
+  ngx-flamegraph {
+    position: relative;
 
-  ::ng-deep {
-    .ngx-fg-label {
-      font-weight: 500;
-      font-size: 1em;
-      color: var(--senary-contrast);
-    }
+    ::ng-deep {
+      .ngx-fg-label {
+        font-weight: 500;
+        font-size: 1em;
+        color: var(--senary-contrast);
+      }
 
-    .ngx-fg-rect {
-      stroke: var(--secondary-contrast);
-      transition: none;
-    }
+      .ngx-fg-rect {
+        stroke: var(--secondary-contrast);
+        transition: none;
+      }
 
-    .ngx-fg-svg-g {
-      transition: none;
+      .ngx-fg-svg-g {
+        transition: none;
+      }
     }
   }
 }
