+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/recording-visualizer/recording-visualizer.component.html
@@ -16,29 +16,26 @@
       }
     }
   </as-split-area>
+
   @if (selectedEntry(); as entry) {
-    <as-split-area size="25" minSize="15">
-      <mat-card class="selected-entry">
-        <mat-toolbar>{{ entry.label }} details</mat-toolbar>
-        <section class="entry-statistics">
-          <div class="txt-total-time">
-            <label>Total time spent:</label><span> {{ entry.value | number }} ms</span>
-          </div>
-          @if (entry.value > 0) {
-            <ng-execution-details [data]="selectedDirectives()" />
-          }
-          @if (parentHierarchy().length > 0) {
-            <div>
-              <div class="txt-total-time"><label>Parent Hierarchy</label></div>
-              <ul>
-                @for (parent of parentHierarchy(); track $index) {
-                  <li>{{ parent.name }}</li>
-                }
-              </ul>
-            </div>
-          }
-        </section>
-      </mat-card>
+    <as-split-area size="25" minSize="15" class="selected-entry">
+      <p class="title">{{ entry.label }} details</p>
+      <div class="entry-statistics">
+        <p class="total-time">
+          Total time spent: <span>{{ entry.value | number }} ms</span>
+        </p>
+        @if (entry.value > 0) {
+          <ng-execution-details [data]="selectedDirectives()" />
+        }
+        @if (parentHierarchy().length > 0) {
+          <p class="subtitle">Parent Hierarchy</p>
+          <ul>
+            @for (parent of parentHierarchy(); track $index) {
+              <li>{{ parent.name }}</li>
+            }
+          </ul>
+        }
+      </div>
     </as-split-area>
   }
 </as-split>
