+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/recording-visualizer/recording-visualizer.component.ts
@@ -7,13 +7,11 @@
  */
 
 import {ChangeDetectionStrategy, Component, computed, input, linkedSignal} from '@angular/core';
-import {MatToolbar} from '@angular/material/toolbar';
-import {MatCard} from '@angular/material/card';
 import {DecimalPipe} from '@angular/common';
 
 import {ProfilerFrame} from '../../../../../../../protocol';
 
-import {VisualizationMode} from '../visualization-mode';
+import {VisualizationMode} from '../shared/visualization-mode';
 import {SelectedEntry} from './recording-visualizer-types';
 
 import {ExecutionDetailsComponent} from './execution-details/execution-details.component';
@@ -34,8 +32,6 @@ import {SplitComponent} from '../../../../vendor/angular-split/lib/component/spl
     FlamegraphVisualizerComponent,
     TreeMapVisualizerComponent,
     BargraphVisualizerComponent,
-    MatCard,
-    MatToolbar,
     ExecutionDetailsComponent,
     DecimalPipe,
   ],
