+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/recording-visualizer/tree-map-visualizer/tree-map-visualizer.component.scss
@@ -1,11 +1,5 @@
 @use '../../../../../../styles/typography';
 
-.web-tree {
-  height: calc(100% - 1.5rem);
-  width: calc(100% - 1.5rem);
-  margin: auto;
-}
-
 :host {
   ::ng-deep {
     .webtreemap-node {
@@ -17,5 +11,12 @@
     .webtreemap-node:hover {
       background: var(--dynamic-blue-01);
     }
+  }
+
+  .web-tree {
+    height: calc(100% - 1.5rem);
+    width: calc(100% - 1.5rem);
+    margin: auto;
+    margin-top: 0.75rem;
   }
 }
