+++ devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/shared/visualization-mode.ts
