--- devtools/projects/ng-devtools/src/lib/devtools-tabs/profiler/recording-timeline/visualization-mode/BUILD.bazel
@@ -1,8 +1,0 @@
-load("//devtools/tools:typescript.bzl", "ts_project")
-
-package(default_visibility = ["//devtools:__subpackages__"])
-
-ts_project(
-    name = "visualization-mode",
-    srcs = ["index.ts"],
-)
